requirements.txt
djangorestframework
cerberus
pytest
